requirements.txt
attrs==16.3.0
scipy==0.19.0
numpy==1.12.1
root-numpy==4.7.1
Shapely==1.5.17.post1
descartes==1.1.0
matplotlib==2.0.0
